fix: 깨지는 테스트코드 수정 및 테스트코드 중복 제거를 위한 상속구조 적용
