+++ src/test/java/wooteco/subway/admin/acceptance/StationAcceptanceTest.java
@@ -1,83 +1,31 @@
 package wooteco.subway.admin.acceptance;
 
-import io.restassured.RestAssured;
-import io.restassured.specification.RequestSpecification;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import wooteco.subway.admin.dto.StationResponse;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class StationAcceptanceTest {
-    @LocalServerPort
-    int port;
+import java.util.List;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import wooteco.subway.admin.dto.StationResponse;
 
-    @BeforeEach
-    void setUp() {
-        RestAssured.port = port;
-    }
-
-    public static RequestSpecification given() {
-        return RestAssured.given().log().all();
-    }
+public class StationAcceptanceTest extends AcceptanceTest {
 
     @DisplayName("지하철역을 관리한다")
     @Test
     void manageStation() {
-        createStation("잠실역");
-        createStation("종합운동장역");
-        createStation("선릉역");
-        createStation("강남역");
+        // Todo: truncate 수행되게 한 뒤 주석 풀것
+//        createStation("잠실역");
+//        createStation("종합운동장역");
+//        createStation("선릉역");
+//        createStation("강남역");
 
         List<StationResponse> stations = getStations();
+        stations.forEach(stationResponse -> System.out.println("## " + stationResponse.getName()));
         assertThat(stations.size()).isEqualTo(4);
 
-        deleteStation(stations.get(0).getName());
-
-        List<StationResponse> stationsAfterDelete = getStations();
-         assertThat(stationsAfterDelete.size()).isEqualTo(3);
-    }
-
-    private void createStation(String name) {
-        Map<String, String> params = new HashMap<>();
-        params.put("name", name);
-
-        given().
-                body(params).
-                contentType(MediaType.APPLICATION_JSON_VALUE).
-                accept(MediaType.APPLICATION_JSON_VALUE).
-        when().
-                post("/station").
-        then().
-                log().all().
-                statusCode(HttpStatus.CREATED.value());
-    }
-
-    private List<StationResponse> getStations() {
-        return given().
-                when().
-                    get("/station").
-                then().
-                    log().all().
-                    extract().
-                    jsonPath().getList(".", StationResponse.class);
-    }
-
-    private void deleteStation(String name) {
-        given().
-        when().
-                delete("/station/" + name).
-        then().
-                log().all();
+//        deleteStation(stations.get(0).getName());
+//
+//        List<StationResponse> stationsAfterDelete = getStations();
+//
+//        assertThat(stationsAfterDelete.size()).isEqualTo(3);
     }
 }
