+++ src/test/java/wooteco/subway/admin/domain/LineTest.java
@@ -1,36 +1,35 @@
 package wooteco.subway.admin.domain;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.time.LocalTime;
+import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-
-import java.time.LocalTime;
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 public class LineTest {
     private Line line;
 
     @BeforeEach
     void setUp() {
-        line = new Line(1L, "2호선", LocalTime.of(05, 30), LocalTime.of(22, 30), 5, "test-color");
+        line = new Line("2호선", LocalTime.of(05, 30), LocalTime.of(22, 30), 5, "test-color");
+        line.addLineStation(new LineStation(null, 1L, 10, 10));
         line.addLineStation(new LineStation(1L, 2L, 10, 10));
-        line.addLineStation(new LineStation(2L, 3L, 10, 10));
     }
 
     @Test
     void getLineStations() {
         List<Long> stationIds = line.getSortedStationIds();
 
-        assertThat(stationIds.size()).isEqualTo(3);
+        assertThat(stationIds.size()).isEqualTo(2);
         assertThat(stationIds.get(0)).isEqualTo(1L);
-        assertThat(stationIds.get(2)).isEqualTo(3L);
+        assertThat(stationIds.get(1)).isEqualTo(2L);
     }
 
     @ParameterizedTest
-    @ValueSource(longs = {1L, 2L, 3L})
+    @ValueSource(longs = {1L, 2L})
     void removeLineStation(Long stationId) {
         line.removeLineStationByStationId(stationId);
