+++ src/test/resources/truncate.sql
@@ -1,3 +1,3 @@
 TRUNCATE TABLE STATION;
 TRUNCATE TABLE LINE;
---TRUNCATE TABLE LINE_STATION;
+TRUNCATE TABLE LINE_STATION;
